removed inline styles in 1 place and refactored
