+++ client/src/components/AvatarDOM.scss
