+++ client/src/pages/JoinPage.scss
@@ -1,0 +1,4 @@
+.visualizer {
+    height: 10px;
+    background: #FDEEC8;
+}
