+++ client/src/templates/RoomTemplate.tsx
@@ -38,7 +38,7 @@ const RoomTemplate = ({
           {sideDrawerComponent}
           <Div d="flex" justify="center" m={{ t: "2rem" }}>
             <Button
-            bg="gray200"
+              bg="gray200"
               textColor="medium"
               onClick={() => setShowModal(false)}
             >
